+++ decode.c
@@ -20,45 +20,45 @@
 #include "interp.h"
 
     /* Canonical coding arrays (flogged from encode.c) */
-extern ulong lj_base[];
-extern ulong min_code[];
-extern ulong offset[];
+extern uint32_t lj_base[];
+extern uint32_t min_code[];
+extern uint32_t offset[];
 
-ulong *lut[LUT_SIZE];       /* canonical decode array */
-uint max_cw_len;
-uint min_cw_len;            /* used to start the linear search if lut==NULL */
+uint32_t *lut[LUT_SIZE];       /* canonical decode array */
+uint32_t max_cw_len;
+uint32_t min_cw_len;            /* used to start the linear search if lut==NULL */
 
     /* prototypes */
-int read_header(FILE *in_file, uint mapping[], uint cw_lens[]);
+int32_t read_header(FILE *in_file, uint32_t mapping[], uint32_t cw_lens[]);
 void build_lut(void);
-void decode(FILE *in_file, uint mapping[]);
-extern void build_canonical_arrays(uint cw_lens[], uint);
+void decode(FILE *in_file, uint32_t mapping[]);
+extern void build_canonical_arrays(uint32_t cw_lens[], uint32_t);
 
-int peak_lens_memory;
+int32_t peak_lens_memory;
 /*
 ** Read compressed stdin, output symbols to stdout
 */
 void
 do_decoding(FILE *in_file) {
-    ulong magic;
+    uint32_t magic;
 
     if (START_INPUT(in_file) == EOF) return;      /* empty input file */
 
-    magic = INPUT_ULONG(in_file, sizeof(ulong)*8);
+    magic = INPUT_ULONG(in_file, sizeof(uint32_t)*8);
     if (magic != MAGIC) {
         fprintf(stderr,"Input file was not compressed with shuff.\n");
         return;
     }
 
-    uint *mapping;    /* ordinal sym -> real sym */
-    uint cw_lens[L+1];
+    uint32_t *mapping;    /* ordinal sym -> real sym */
+    uint32_t cw_lens[L+1];
 
-    allocate(mapping, uint, MAX_SYMBOL);
-    SHOW_MEM(L, ulong) /* min_code */
-    SHOW_MEM(L, ulong) /* lj_base */
-    SHOW_MEM(L, ulong) /* offset */
-    SHOW_MEM(L+1, uint) /* cw_lens */
-    SHOW_MEM(MAX_SYMBOL, uint) /* mapping */
+    allocate(mapping, uint32_t, MAX_SYMBOL);
+    SHOW_MEM(L, uint32_t) /* min_code */
+    SHOW_MEM(L, uint32_t) /* lj_base */
+    SHOW_MEM(L, uint32_t) /* offset */
+    SHOW_MEM(L+1, uint32_t) /* cw_lens */
+    SHOW_MEM(MAX_SYMBOL, uint32_t) /* mapping */
     peak_lens_memory = 0;
 
     while (read_header(in_file, mapping, cw_lens) != 0) {
@@ -66,7 +66,7 @@ do_decoding(FILE *in_file) {
         decode(in_file, mapping);
     }
 
-    SHOW_MEM(peak_lens_memory, int) /* mapping */
+    SHOW_MEM(peak_lens_memory, int32_t) /* mapping */
 } /* do_decoding() */
 
 /*
@@ -80,16 +80,16 @@ do_decoding(FILE *in_file) {
 ** (5) Build cw_lens[]
 ** (6) Build mapping[]
 */
-int
-read_header(FILE *in_file, uint mapping[], uint cw_lens[]) {
-    int i, n;
-    int *lens, *p;
+int32_t
+read_header(FILE *in_file, uint32_t mapping[], uint32_t cw_lens[]) {
+    int32_t i, n;
+    int32_t *lens, *p;
 
     n = INPUT_ULONG(in_file, LOG2_MAX_SYMBOL);
 //fprintf(stderr,"n: %10u\n",n);
     if (n == 0) return 0;      /* last block */
 
-    allocate(lens, int, n+1);
+    allocate(lens, int32_t, n+1);
     if (n+1 > peak_lens_memory) 
         peak_lens_memory = n+1;
     
@@ -105,7 +105,7 @@ read_header(FILE *in_file, uint mapping[], uint cw_lens[]) {
 
     for(min_cw_len = 0 ; cw_lens[min_cw_len] == 0 ; min_cw_len++);
 
-//{uint i;
+//{uint32_t i;
 //fprintf(stderr,"cw_lens : \n");
 //for(i=min_cw_len;i<=max_cw_len;i++)
 //fprintf(stderr,"%2d) %u\n",i, cw_lens[i]);
@@ -116,14 +116,14 @@ read_header(FILE *in_file, uint mapping[], uint cw_lens[]) {
 
     interp_decode(in_file, mapping, n);
 
-    for( i = 1 ; i <= (int)max_cw_len; i++)
+    for( i = 1 ; i <= (int32_t)max_cw_len; i++)
         cw_lens[i] += cw_lens[i-1];     
 
     for (p = lens + n - 1 ; p >= lens ; p--)
         *p = cw_lens[*p - 1]++;
 
-    uint t, from, S;
-    int  start = 0;
+    uint32_t t, from, S;
+    int32_t  start = 0;
     lens[n] = 1;            // sentinel
     while (start < n) {
         from = start;
@@ -139,7 +139,7 @@ read_header(FILE *in_file, uint mapping[], uint cw_lens[]) {
         while (lens[start] == -1) start++; // find next start (if any)
     }
 
-//{int i;
+//{int32_t i;
 //fprintf(stderr,"mapping\n");
 //for(i = 0 ; i < n; i++)
 //fprintf(stderr,"%8u %8u\n",i, mapping[i]);
@@ -153,11 +153,11 @@ read_header(FILE *in_file, uint mapping[], uint cw_lens[]) {
 
 void
 build_lut() {
-    uint max, min;                // range of left justified "i"
-    int i,j = max_cw_len - 1;  // pointer into lj
+    uint32_t max, min;                // range of left justified "i"
+    int32_t i,j = max_cw_len - 1;  // pointer into lj
 
     for(i = 0; i < LUT_SIZE ; i++) {
-        min = i << ((sizeof(ulong) << 3) - LUT_BITS);
+        min = i << ((sizeof(uint32_t) << 3) - LUT_BITS);
         max = min | MAX_IT;
 
         while ( (j >=0 ) && (max > lj_base[j])) j--;
@@ -168,7 +168,7 @@ build_lut() {
         else
             lut[i] = NULL; //-(j+1);
     }
-//{int i;
+//{int32_t i;
 //for(i = 0 ; i < (1 << LUT_BITS) ; i++)
 //if (lut[i] == NULL)
 //fprintf(stderr, "lut[%2x] NULL\n",i);
@@ -187,28 +187,28 @@ build_lut() {
 ** Decode canonical codes until we get symbol EOF_SYMBOL
 */
 void
-decode(FILE *in_file, uint mapping[]) {
+decode(FILE *in_file, uint32_t mapping[]) {
     #define BUFF_LENGTH 4096
-    ulong buffer[BUFF_LENGTH];
-    ulong *buff = buffer;
+    uint32_t buffer[BUFF_LENGTH];
+    uint32_t *buff = buffer;
 
-    ulong code        = 0;
-    ulong bits_needed = sizeof(ulong) << 3;
-    ulong currcode;
-    ulong currlen     = sizeof(ulong) << 3;
-    ulong *lj;
-    ulong *start_linear_search = lj_base + MAX(LUT_BITS, min_cw_len) - 1;
+    uint32_t code        = 0;
+    uint32_t bits_needed = sizeof(uint32_t) << 3;
+    uint32_t currcode;
+    uint32_t currlen     = sizeof(uint32_t) << 3;
+    uint32_t *lj;
+    uint32_t *start_linear_search = lj_base + MAX(LUT_BITS, min_cw_len) - 1;
 
     for(;;) {
         code |= INPUT_ULONG(in_file, bits_needed);
 
-        lj = lut[code >> ((sizeof(ulong)<<3) - LUT_BITS)];
+        lj = lut[code >> ((sizeof(uint32_t)<<3) - LUT_BITS)];
         if (lj == NULL)
             for(lj = start_linear_search ; code < *lj ; lj++);
         currlen = lj - lj_base + 1;
 
             // calculate symbol number
-        currcode = code >> ((sizeof(ulong) << 3) - currlen);
+        currcode = code >> ((sizeof(uint32_t) << 3) - currlen);
         currcode -= min_code[currlen-1];
         currcode += offset[currlen-1];
 
@@ -219,7 +219,7 @@ decode(FILE *in_file, uint mapping[]) {
     break;
 
         if (buff == buffer + BUFF_LENGTH) {
-            fwrite(buffer,sizeof(ulong),BUFF_LENGTH,stdout);
+            fwrite(buffer,sizeof(uint32_t),BUFF_LENGTH,stdout);
             buff = buffer;
         }
             // subtract the one added in encoding
@@ -231,9 +231,9 @@ decode(FILE *in_file, uint mapping[]) {
     }
 
     if (buff == buffer + BUFF_LENGTH) {
-        fwrite(buffer,sizeof(ulong),BUFF_LENGTH,stdout);
+        fwrite(buffer,sizeof(uint32_t),BUFF_LENGTH,stdout);
         buff = buffer;
     }
-    fwrite(buffer,sizeof(ulong),buff - buffer,stdout);
+    fwrite(buffer,sizeof(uint32_t),buff - buffer,stdout);
 
 } /* decode() */
