+++ nqsort.c
@@ -1,3 +1,5 @@
+#include <stdint.h>
+
 /*
  * Copyright (c) 1992 The Regents of the University of California.
  * All rights reserved.
@@ -8,7 +10,7 @@
  * Qsort routine from J. Bentley & M. D. McIlroy's "Engineering a Sort Function".
  */
 #define swapcode(TYPE, parmi, parmj, n) { \
-	long i = (n) / sizeof (TYPE); \
+	int32_t i = (n) / sizeof (TYPE); \
 	register TYPE *pi = (TYPE *) (parmi); \
 	register TYPE *pj = (TYPE *) (parmj); \
 	do { \
@@ -18,29 +20,29 @@
         } while (--i > 0);	\
 }
 
-#define SWAPINIT(a, es) swaptype = ((char *)a - (char *)0) % sizeof(long) || \
-	es % sizeof(long) ? 2 : es == sizeof(long)? 0 : 1;
+#define SWAPINIT(a, es) swaptype = ((char *)a - (char *)0) % sizeof(int32_t) || \
+	es % sizeof(int32_t) ? 2 : es == sizeof(int32_t)? 0 : 1;
 
 void
-swapfunc(char *a, char *b, int n, int swaptype) {
+swapfunc(char *a, char *b, int32_t n, int32_t swaptype) {
 	if(swaptype <= 1) 
-		swapcode(long, a, b, n)
+		swapcode(int32_t, a, b, n)
 	else
 		swapcode(char, a, b, n)
 }
 
 #define swap(a, b) \
 	if (swaptype == 0) { \
-		long t = *(long *)(a); \
-		*(long *)(a) = *(long *)(b); \
-		*(long *)(b) = t; \
+		int32_t t = *(int32_t *)(a); \
+		*(int32_t *)(a) = *(int32_t *)(b); \
+		*(int32_t *)(b) = t; \
 	} else \
 		swapfunc(a, b, es, swaptype)
 
 #define vecswap(a, b, n) 	if ((n) > 0) swapfunc(a, b, n, swaptype)
 
 char *
-med3(char *a, char *b, char *c, int (*cmp)(char * , char *)) {
+med3(char *a, char *b, char *c, int32_t (*cmp)(char * , char *)) {
 	return cmp(a, b) < 0 ?
 	       (cmp(b, c) < 0 ? b : (cmp(a, c) < 0 ? c : a ))
               :(cmp(b, c) > 0 ? b : (cmp(a, c) < 0 ? a : c ));
@@ -49,9 +51,9 @@ med3(char *a, char *b, char *c, int (*cmp)(char * , char *)) {
 #define min(a, b) (a) < (b) ? a : b
 
 void
-nqsort(char *a, unsigned int n, int es, int (*cmp)(char *, char *)) {
+nqsort(char *a, uint32_t n, int32_t es, int32_t (*cmp)(char *, char *)) {
 	char *pa, *pb, *pc, *pd, *pl, *pm, *pn;
-	int d, r, swaptype;
+	int32_t d, r, swaptype;
 
 	SWAPINIT(a, es);
 	if (n < 7) {
